More sbnet progress and general bugfixes
This adds the ability to generically decode packets to sbnet, which is
nice. There's definitely some sort of race or memory corruption causing
issues in the code base but we can track that down once we have this
working. This fixes a couple minor bugs in the scheduler as well