Minor refactoring and remove auto-generated code
Added a base plugin struct, right now it's just a lock. Adapted the net
and event plugin to use it. Added some comments and other small
re-factor stuff throughout the codebase.

Changed the way the code base is built, protocol files are now generated
by cmake instead of being carried in commits. This now requires mcd2c be
installed to pythonpath which complicates things slightly, need to make
the on-boarding process easier for people eventually.